requirements.txt
streamlit
Pillow
torch
torchvision
tensorflow
numpy
